(GEN-141) Catch errors from remote invocation of go-functions
A Go function has an option to return an error. When it does,
the remote invocation logic will treat this as a controlled
way of propagating the error to the client. The service will
hence not panic but instead return an Error object to the
client.

This commit ensures that the Error object in turn, causes a
panic with an issue.Reported object so that the error isn't
propagated all the way up to the caller where an Error isn't
expected.