Improve TypeScript generator to generate interfaces for anonymous types
Since the anonymous types need to be repeated both in the attribute
declaration and in the parameter type declaration, it saves many lines
of codes to instead generate an interface. Here, interfaces are
generated with the name "Anon" followed by a sequence number. The user
will never have to write those names.