Add special NotFound() function to produce well known error.
This commit adds a NotFound() function which returns an error. The
function should be used by providers when a request has been made to
read or update a non existing resource using an external id. This
condition is normally ok and indicates that the identity storage
contains references to resources that no longer exists. By returning
the NotFound error, the workflow engine will be made aware of this and
purge the external ID from the DB.

Relates to lyraproj/lyra#214