Only create definitions from APIs that matters to the workflow
This commit ensures that only those APIs that matters to the workflow
engine are propagated as Definitions. Other APIs (such as the function
of an Action) should not be propagated since they cause collisions in
the definition namespace.

This replaces an earlier failed attempt to add an '::Api' suffix to
APIs irrelevant to the wf-engine.