Ensure correct serialization of error arguments to issue.Reported
If an argument to an issue.Reported is of type `error` it must be
converted to a string using the Error() method before being serialized
to a stream because there's no way of knowing what the underlying
go type might be or how it will be serialized.

This commit ensures that the string conversion takes place before the
argument is stored in the ErrorObject that is used when the error
is serialized/deserialized.

Closes lyraproj/lyra#292