Improve error reporting of errors propagated over grpc
This commit ensures that errors that are propagated over grpc contains
information about host, executable, and optionally stack (depends on if
debug logging is active or not). Errors are then reported with this
information unless the caller's information is the same (i.e. host
is not reported unless call is to another host).