Let methods of Identity service take px.Context argument
In order for the client side of the Identity service to use the same
interface as the server, the interface methods must take a px.Context
as their first parameter. This commit fixes that and also tweaks the
methods GetInternal and GetExternal so that they return the result
together with a boolean.